v0.2.0 - region tuples, region-get result change
Bumping version for: changing result type of region-gets.

This backwards incompatibility should be a helpful change.
Given a region, we now return the smallest possible integer type that
can hold the region (`u8`, `u16, `u32`, `u64`, or `usize` as last
resort).

Also introduces getting regions as a tuple of `u8` or `bool` for easy
pattern-matching.